Php8.2 Deprecation in Reader/Xlsx (#2894)
* Php8.2 Deprecation in Reader/Xlsx

Using `${var}` will be deprecated, with the suggested resolution being to use `{$var}`. This appears to be the only place in PhpSpreadsheet which does this. Some vendor packages will need to change for 8.2 for this and other reasons.

* mb_convert_encoding and HTML_ENTITIES

Also scheduled to be deprecated with 8.2. It appears to have not been needed in PhpSpreadsheet in the first place.